+++ GWToolbox/GWToolbox/OtherModules/Resources.cpp
@@ -14,11 +14,11 @@ void Resources::Initialize() {
 	ToolboxModule::Initialize();
 	worker = std::thread([this]() {
 		while (!should_stop) {
-			if (todo.empty()) {
+			if (thread_jobs.empty()) {
 				std::this_thread::sleep_for(std::chrono::milliseconds(100));
 			} else {
-				todo.front()();
-				todo.pop();
+				thread_jobs.front()();
+				thread_jobs.pop();
 			}
 		}
 	});
@@ -28,94 +28,98 @@ void Resources::Terminate() {
 	if (worker.joinable()) worker.join();
 }
 void Resources::EndLoading() {
-	todo.push([this]() { should_stop = true; });
+	thread_jobs.push([this]() { should_stop = true; });
 }
 
-bool Resources::GetPath(CHAR* path, const char* folder) const {
-	if (!SUCCEEDED(SHGetFolderPath(NULL, CSIDL_LOCAL_APPDATA | CSIDL_FLAG_CREATE, NULL, 0, path))) return false;
-	PathAppend(path, "GWToolboxpp");
-	if (folder) PathAppend(path, folder);
-	return true;
-}
-
-bool Resources::GetURL(CHAR* url, const char* name, const char* folder) const {
-	// todo: find a better way that doesn't break if folder has a trailing slash
-	if (folder) {
-		sprintf_s(url, MAX_PATH, "%s%s/%s", GWTOOLBOX_HOST, folder, name);
-	} else {
-		sprintf_s(url, MAX_PATH, "%s%s", GWTOOLBOX_HOST, name);
-	}
-	return true;
-}
-
-void Resources::EnsureFullPathExists(const char* path) const {
-	if (!PathFileExists(path)) {
-		CreateDirectory(path, NULL);
-	}
-}
-
-void Resources::EnsureSubPathExists(const char* sub) const {
+std::string Resources::GetSettingsFolderPath() {
 	CHAR path[MAX_PATH];
-	if (GetPath(path, sub)) {
-		EnsureFullPathExists(path);
+	SHGetFolderPath(NULL, CSIDL_LOCAL_APPDATA | CSIDL_FLAG_CREATE, NULL, 0, path);
+	PathAppend(path, "GWToolboxpp");
+	return std::string(path);
+
+}
+std::string Resources::GetPath(std::string file) {
+	return GetSettingsFolderPath() + "\\" + file;
+}
+std::string Resources::GetPath(std::string folder, std::string file) {
+	return GetSettingsFolderPath() + "\\" + folder + "\\" + file;
+}
+
+void Resources::EnsureFolderExists(std::string path) {
+	if (!PathFileExists(path.c_str())) {
+		CreateDirectory(path.c_str(), NULL);
 	}
 }
 
-void Resources::LoadTextureAsync(IDirect3DTexture9** texture, const char* name, const char* folder) {
-	CHAR* path = new CHAR[MAX_PATH];
-	if (!GetPath(path, folder)) return;
-	EnsureFullPathExists(path);
-	PathAppend(path, name);
+bool Resources::Download(std::string path_to_file, std::string url) const {
+	DeleteUrlCacheEntry(url.c_str());
+	Log::Log("Downloading %s\n", url.c_str());
+	return (URLDownloadToFile(NULL, url.c_str(), path_to_file.c_str(), 0, NULL) == S_OK);
+}
+void Resources::Download(std::string path_to_file, 
+	std::string url, std::function<void(bool)> callback) {
 
-	if (PathFileExists(path)) {
-		toload.push([path, texture](IDirect3DDevice9* device) {
-			D3DXCreateTextureFromFile(device, path, texture);
-			delete[] path;
+	thread_jobs.push([this, url, path_to_file, callback]() {
+		bool success = Download(path_to_file, url);
+		// and call the callback in the main thread
+		todo.push([callback, success]() { callback(success); });
+		if (!success) {
+			Log::Log("Error downloading %s\n", url.c_str());
+		}
+	});
+}
+
+std::string Resources::Download(std::string url) const {
+	DeleteUrlCacheEntry(url.c_str());
+	IStream* stream;
+	std::string ret = "";
+	if (SUCCEEDED(URLOpenBlockingStream(NULL, url.c_str(), &stream, 0, NULL))) {
+		STATSTG stats;
+		stream->Stat(&stats, STATFLAG_NONAME);
+		DWORD size = stats.cbSize.LowPart;
+		CHAR* chars = new CHAR[size + 1];
+		stream->Read(chars, size, NULL);
+		chars[size] = '\0';
+		ret = std::string(chars);
+		delete[] chars;
+		stream->Release();
+	}
+	return ret;
+}
+void Resources::Download(std::string url, std::function<void(std::string)> callback) {
+	thread_jobs.push([this, url, callback]() {
+		std::string s = Download(url);
+		todo.push([callback, s]() { callback(s); });
+	});
+}
+
+void Resources::EnsureFileExists(std::string path_to_file, 
+	std::string url, std::function<void(bool)> callback) {
+
+	if (PathFileExists(path_to_file.c_str())) {
+		// if file exists, run the callback immediately in the same thread
+		callback(true);
+	} else {
+		// otherwise try to download it in the worker
+		Download(path_to_file, url, callback);
+	}
+}
+
+void Resources::LoadTextureAsync(IDirect3DTexture9** texture, 
+	std::string path_to_file, std::string url) {
+
+	if (PathFileExists(path_to_file.c_str())) {
+		toload.push([path_to_file, texture](IDirect3DDevice9* device) {
+			D3DXCreateTextureFromFile(device, path_to_file.c_str(), texture);
 		});
 	} else {
-		CHAR* url = new CHAR[MAX_PATH];
-		GetURL(url, name, folder);
-		todo.push([this, url, path, texture]() {
-			DeleteUrlCacheEntry(url);
-			Log::Log("Downloading %s\n", url);
-			if (URLDownloadToFile(NULL, url, path, 0, NULL) == S_OK) {
-				toload.push([path, texture](IDirect3DDevice9* device) {
-					D3DXCreateTextureFromFile(device, path, texture);
-					delete[] path;
+		Download(path_to_file, url, 
+			[this, path_to_file, texture](bool success) {
+			if (success) {
+				toload.push([path_to_file, texture](IDirect3DDevice9* device) {
+					D3DXCreateTextureFromFile(device, path_to_file.c_str(), texture);
 				});
-			} else {
-				delete[] path;
-				Log::Log("Error downloading %s\n", url);
 			}
-			delete[] url;
-		});
-	}
-}
-
-void Resources::EnsureFileExists(const char* name, 
-	const char* folder, std::function<void()> callback) {
-
-	CHAR* path = new CHAR[MAX_PATH];
-	if (!GetPath(path, folder)) return;
-	EnsureFullPathExists(path);
-	PathAppend(path, name);
-
-	if (PathFileExists(path)) {
-		callback();
-		delete[] path;
-	} else {
-		CHAR* url = new CHAR[MAX_PATH];
-		GetURL(url, name, folder);
-		todo.push([this, url, path, callback]() {
-			DeleteUrlCacheEntry(url);
-			Log::Log("Downloading %s\n", url);
-			if (URLDownloadToFile(NULL, url, path, 0, NULL) == S_OK) {
-				callback();
-			} else {
-				Log::Log("Error downloading %s\n", url);
-			}
-			delete[] url;
-			delete[] path;
 		});
 	}
 }
@@ -124,5 +128,12 @@ void Resources::DxUpdate(IDirect3DDevice9* device) {
 	while (!toload.empty()) {
 		toload.front()(device);
 		toload.pop();
+	}
+}
+
+void Resources::Update() {
+	while (!todo.empty()) {
+		todo.front()();
+		todo.pop();
 	}
 }