+++ GWToolbox/GWToolbox/Panels/Hotkeys.h
@@ -23,6 +23,7 @@ public:
 
 	static bool show_active_in_header;
 	static bool show_run_in_header;
+	static bool hotkeys_changed;
 
 	static TBHotkey* HotkeyFactory(CSimpleIni* ini, const char* section);