+++ GWToolbox/GWToolbox/Windows/PartyDamage.cpp
@@ -10,6 +10,7 @@
 
 #include <GWToolbox.h>
 #include <GuiUtils.h>
+#include <OtherModules\Resources.h>
 #include <OtherModules\ToolboxSettings.h>
 
 #define IniFilename "healthlog.ini"
@@ -337,7 +338,7 @@ void PartyDamage::LoadSettings(CSimpleIni* ini) {
 	color_recent = Colors::Load(ini, Name(), "color_recent", Colors::ARGB(205, 102, 153, 230));
 
 	if (inifile == nullptr) inifile = new CSimpleIni(false, false, false);
-	inifile->LoadFile(GuiUtils::getPath(IniFilename).c_str());
+	inifile->LoadFile(Resources::GetPath(IniFilename).c_str());
 	CSimpleIni::TNamesDepend keys;
 	inifile->GetAllKeys(IniSection, keys);
 	for (CSimpleIni::Entry key : keys) {
@@ -367,7 +368,7 @@ void PartyDamage::SaveSettings(CSimpleIni* ini) {
 		std::string key = std::to_string(item.first);
 		inifile->SetLongValue(IniSection, key.c_str(), item.second, 0, false, true);
 	}
-	inifile->SaveFile(GuiUtils::getPath(IniFilename).c_str());
+	inifile->SaveFile(Resources::GetPath(IniFilename).c_str());
 }
 
 void PartyDamage::DrawSettingInternal() {